Convert SidebarInjector to TypeScript
Convert sidebar-injector.js to TS and do a minimal class-ification in the
process. There is further work required to make it a conventional ES class (ie.
without all the code in the constructor).
